requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
packaging==24.0
pillow==10.2.0
PyJWT==2.8.0
python-http-client==3.3.7
requests==2.31.0
sendgrid==3.6.5
sendgrid-django==4.2.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
django_celery_beat
